+++ ios/Classes/LocationController.swift
@@ -102,6 +102,10 @@ extension LocationOptions {
                 configuration.pulsing = pulsing
             }
 
+            if let opacity = puck2D.opacity?.doubleValue {
+                configuration.opacity = opacity
+            }
+
             options.puckType = .puck2D(configuration)
         }
         return options
@@ -141,6 +145,7 @@ extension LocationOptions {
                 let encoded = try! JSONEncoder().encode(scaleData)
                 locationPuck2D.scaleExpression = String(data: encoded, encoding: .utf8)
             }
+            locationPuck2D.opacity = NSNumber(value: oldConfiguration.opacity)
             if let pulsing = oldConfiguration.pulsing {
                 pulsingEnabled = NSNumber(value: true)
                 switch pulsing.radius {